Split `test_task_api.py` into two modules
